+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.16
 
 * changed semantics of `wrap_future`
+* added `Directory` to schema configuration